+++ Changelog.md
@@ -1,4 +1,14 @@
 # Changelog
------------
+
 08-02-2021
-* list_virtual_centers
+* renamed Settings class to Config
+* Config class
+    * Added
+        * list_virtual_centers
+        * get_environment_properties
+        * get_settings
+        * get_settings-feature
+        * get_settings-general
+        * get_settings-security
+    * Added but not working yet
+        * update_settings_general